Adress zero index tl issue (#41)
Adressing waypoint zero index traffic light issue

When the car reaches end of the track in the simulator
waypoint index resets to 0. That would cause problems for
traffic light detection before this fix because
traffic lights are detected based on the difference
between waypoint index of car nad waypoint index of traffic
light.